# apps/derrops-cloud/src/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
    "/services": {
      "post": {
        "operationId": "ServiceController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateServiceDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Service created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Service"
                }
              }
            }
          },
          "400": {
            "description": "Bad request"
          }
        },
        "summary": "Create a new service",
        "tags": [
          "Service"
        ]
      },
      "get": {
        "operationId": "ServiceController_findAll",
        "parameters": [
          {
            "name": "select",
            "required": false,
            "in": "query",
            "description": "Comma-separated list of fields to select (e.g., id,name,endpoint)",
            "schema": {
              "example": "id,name,endpoint,openapi_doc_url,openapi_doc_content",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of services",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Service"
                  }
                }
              }
            }
          }
        },
        "summary": "Get all services",
        "tags": [
          "Service"
        ]
      }
    },
    "/services/{id}": {
      "get": {
        "operationId": "ServiceController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Service UUID",
            "schema": {
              "example": "5c963787-d89d-4260-adaf-6541c41cb982",
              "type": "string"
            }
          },
          {
            "name": "select",
            "required": false,
            "in": "query",
            "description": "Comma-separated list of fields to select",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Service found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Service"
                }
              }
            }
          },
          "404": {
            "description": "Service not found"
          }
        },
        "summary": "Get a service by ID",
        "tags": [
          "Service"
        ]
      },
      "patch": {
        "operationId": "ServiceController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Service UUID",
            "schema": {
              "example": "5c963787-d89d-4260-adaf-6541c41cb982",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateServiceDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Service updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Service"
                }
              }
            }
          },
          "400": {
            "description": "Bad request"
          },
          "404": {
            "description": "Service not found"
          }
        },
        "summary": "Update a service",
        "tags": [
          "Service"
        ]
      },
      "delete": {
        "operationId": "ServiceController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Service UUID",
            "schema": {
              "example": "5c963787-d89d-4260-adaf-6541c41cb982",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Service deleted successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Service deleted successfully"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Service not found"
          }
        },
        "summary": "Delete a service",
        "tags": [
          "Service"
        ]
      }
    },
...
    "/cloud-relay/relay-instance": {
      "get": {
        "operationId": "RelayInstanceController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RelayInstance"
                  }
                }
              }
            }
          }
        },
        "summary": "List all relay instances for the tenant",
        "tags": [
          "Relay Instance"
        ]
      },
      "post": {
        "operationId": "RelayInstanceController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRelayInstanceDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RelayInstance"
                }
              }
            }
          }
        },
        "summary": "Register a new relay instance",
        "tags": [
          "Relay Instance"
        ]
      }
    },
    "/cloud-relay/relay-instance/{id}": {
      "get": {
        "operationId": "RelayInstanceController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "RelayInstance UUID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RelayInstance"
                }
              }
            }
          },
          "404": {
            "description": ""
          }
        },
        "summary": "Get a relay instance by ID",
        "tags": [
          "Relay Instance"
        ]
      },
      "patch": {
        "operationId": "RelayInstanceController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "RelayInstance UUID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateRelayInstanceDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RelayInstance"
                }
              }
            }
          },
          "404": {
            "description": ""
          }
        },
        "summary": "Update a relay instance",
        "tags": [
          "Relay Instance"
        ]
      },
      "delete": {
        "operationId": "RelayInstanceController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "RelayInstance UUID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          },
          "404": {
            "description": ""
          }
        },
        "summary": "Delete a relay instance",
        "tags": [
          "Relay Instance"
        ]
      }
    },
...
  },
  "info": {
    "title": "Derrops API",
    "description": "Internal API",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "CreateServiceDto": {
        "type": "object",
        "properties": {
          "user_id": {
            "type": "string",
            "description": "User ID who owns this service",
            "example": "5c963787-d89d-4260-adaf-6541c41cb982"
          },
          "name": {
            "type": "string",
            "description": "Service name",
            "example": "SendGrid API"
          },
          "endpoint": {
            "type": "string",
            "format": "uri",
            "description": "Service endpoint URL",
            "example": "https://api.sendgrid.com/v3"
          },
          "openapi_doc_url": {
            "type": "string",
            "nullable": true,
            "description": "OpenAPI document URL",
            "example": "https://raw.githubusercontent.com/sendgrid/sendgrid-oai/main/oai.json"
          },
          "openapi_doc_content": {
            "type": "string",
            "nullable": true,
            "description": "OpenAPI document content (stored as text)"
          },
          "availability": {
            "type": "number",
            "nullable": true,
            "minimum": 0,
            "maximum": 100,
            "description": "Service availability percentage",
            "example": 99.98
          },
          "response_time": {
            "type": "number",
            "nullable": true,
            "minimum": 0,
            "description": "Average response time in milliseconds",
            "example": 80
          }
        },
        "required": [
          "user_id",
          "name",
          "endpoint"
        ]
      },
      "Service": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Primary key in database",
            "example": "5c963787-d89d-4260-adaf-6541c41cb982"
          },
          "user_id": {
            "type": "string",
            "description": "User ID who owns this service",
            "example": "5c963787-d89d-4260-adaf-6541c41cb982"
          },
          "name": {
            "type": "string",
            "description": "Service name",
            "example": "SendGrid API"
          },
          "endpoint": {
            "type": "string",
            "description": "Service endpoint URL",
            "example": "https://api.sendgrid.com/v3"
          },
          "openapi_doc_url": {
            "type": "string",
            "nullable": true,
            "description": "OpenAPI document URL",
            "example": "https://raw.githubusercontent.com/sendgrid/sendgrid-oai/main/oai.json"
          },
          "openapi_doc_content": {
            "type": "string",
            "nullable": true,
            "description": "OpenAPI document content (stored as text)"
          },
          "availability": {
            "type": "number",
            "nullable": true,
            "description": "Service availability percentage",
            "example": 99.98,
            "minimum": 0,
            "maximum": 100
          },
          "response_time": {
            "type": "number",
            "nullable": true,
            "description": "Average response time in milliseconds",
            "example": 80,
            "minimum": 0
          },
          "created_at": {
            "format": "date-time",
            "type": "string",
            "description": "Date when the service was created",
            "example": "2024-01-15T10:30:00.000Z"
          },
          "updated_at": {
            "format": "date-time",
            "type": "string",
            "description": "Date when the service was last updated",
            "example": "2024-01-15T10:30:00.000Z"
          }
        },
        "required": [
          "id",
          "user_id",
          "name",
          "endpoint",
          "created_at",
          "updated_at"
        ]
      },
      "UpdateServiceDto": {
        "type": "object",
        "properties": {
          "user_id": {
            "type": "string",
            "description": "User ID who owns this service",
            "example": "5c963787-d89d-4260-adaf-6541c41cb982"
          },
          "name": {
            "type": "string",
            "description": "Service name",
            "example": "SendGrid API"
          },
          "endpoint": {
            "type": "string",
            "format": "uri",
            "description": "Service endpoint URL",
            "example": "https://api.sendgrid.com/v3"
          },
          "openapi_doc_url": {
            "type": "string",
            "nullable": true,
            "description": "OpenAPI document URL",
            "example": "https://raw.githubusercontent.com/sendgrid/sendgrid-oai/main/oai.json"
          },
          "openapi_doc_content": {
            "type": "string",
            "nullable": true,
            "description": "OpenAPI document content (stored as text)"
          },
          "availability": {
            "type": "number",
            "nullable": true,
            "minimum": 0,
            "maximum": 100,
            "description": "Service availability percentage",
            "example": 99.98
          },
          "response_time": {
            "type": "number",
            "nullable": true,
            "minimum": 0,
            "description": "Average response time in milliseconds",
            "example": 80
          }
        }
      },
...
      "RelayInstance": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "tenant_id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "url": {
            "type": "string"
          },
          "aegis_id": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "object"
          },
          "last_seen_at": {
            "format": "date-time",
            "type": "string",
            "nullable": true
          },
          "created_at": {
            "format": "date-time",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "tenant_id",
          "name",
          "url",
          "aegis_id",
          "status",
          "last_seen_at",
          "created_at",
          "updated_at"
        ]
      },
      "CreateRelayInstanceDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 255,
            "description": "Human-readable name for this relay instance",
            "example": "prod-relay-eu"
          },
          "url": {
            "type": "string",
            "format": "uri",
            "description": "Base URL of the relay (must be HTTPS)",
            "example": "https://relay.example.com"
          }
        },
        "required": [
          "name",
          "url"
        ]
      },
      "UpdateRelayInstanceDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 255,
            "description": "Human-readable name for this relay instance",
            "example": "prod-relay-eu"
          },
          "url": {
            "type": "string",
            "format": "uri",
            "description": "Base URL of the relay (must be HTTPS)",
            "example": "https://relay.example.com"
          },
          "aegisId": {
            "type": "string",
            "description": "UUID of the linked Aegis instance (must belong to the same tenant)"
          }
        }
      },
...
    }
  }
}